project/requirements.txt
fastapi==0.115.4
uvicorn==0.22.0
sqlalchemy[asyncio]==2.0.15
psycopg[binary]==3.2.3
pydantic==1.10.7
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.0.1
python-decouple==3.8